Make sure there is no use after free due to Detour only changing address once transaction is commited.
For now we use a temporary vector for ordinals, but expose the detour pointer to ExtraPatchActions. This is useful if you want to just wrap (detour) a function and be able to call its original version.
However, this means that the user provided ExtraPatchAction must be alive until the end of the patching.
This involves a breaking change in the API for